# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Magic Program",
        "description": "API documentation for Magic Program",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/entry-groups": {
            "get": {
                "tags": [
                    "Entry Groups"
                ],
                "summary": "List all entry groups",
                "description": "Returns a paginated list of entry groups including entry and entry-type counts.",
                "operationId": "getEntryGroups",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Results per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Column to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EntryGroup"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Meta"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Links"
                                            }
                                        },
                                        "pagination_info": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaginationInfo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Entry Groups"
                ],
                "summary": "Create a new entry group",
                "operationId": "createEntryGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "handle",
                                    "status_group_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Human-readable name",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Blog Posts"
                                    },
                                    "handle": {
                                        "description": "URL-safe identifier (must be unique)",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "blog-posts"
                                    },
                                    "description": {
                                        "description": "Optional description",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "sort_order": {
                                        "description": "Display sort position",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "status_group_id": {
                                        "description": "ID of the status group to use",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "field_layout_id": {
                                        "description": "ID of an optional field layout",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "category_groups": {
                                        "description": "IDs of category groups to attach",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    },
                                    "field_groups": {
                                        "description": "IDs of field groups to attach",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entry group created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/entry-groups/{group}": {
            "get": {
                "tags": [
                    "Entry Groups"
                ],
                "summary": "Get a single entry group",
                "operationId": "getEntryGroup",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "ID of the entry group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Entry Groups"
                ],
                "summary": "Update an entry group",
                "operationId": "updateEntryGroup",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "ID of the entry group to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "handle",
                                    "status_group_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Blog Posts"
                                    },
                                    "handle": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "blog-posts"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "status_group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "field_layout_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "category_groups": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    },
                                    "field_groups": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entry group updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Entry Groups"
                ],
                "summary": "Delete an entry group",
                "description": "Permanently deletes the entry group and all of its entries.",
                "operationId": "deleteEntryGroup",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "ID of the entry group to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Meta": {
                "title": "Collection Meta",
                "description": "",
                "properties": {
                    "current_page": {
                        "description": "The number of the page the results arise from",
                        "type": "integer",
                        "format": "int64"
                    },
                    "from": {
                        "description": "",
                        "type": "integer",
                        "format": "int64"
                    },
                    "last_page": {
                        "description": "The last page number for results",
                        "type": "integer",
                        "format": "int64"
                    },
                    "path": {
                        "description": "The full URL to the given results",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "How many results each request will contain",
                        "type": "number",
                        "format": "float"
                    },
                    "to": {
                        "description": "",
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "description": "The total results available",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Links": {
                "title": "Collection Links",
                "description": "Contains the links used to paginate ",
                "properties": {
                    "first": {
                        "description": "The full URL to the first page of results",
                        "type": "string"
                    },
                    "last": {
                        "description": "The full URL to the last page of results",
                        "type": "string"
                    },
                    "prev": {
                        "description": "The full URL to the prev page of results",
                        "type": "string"
                    },
                    "next": {
                        "description": "The full URL to the next page of results",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginationInfo": {
                "title": "Collection Pagination Info",
                "description": "Minor details about pagination",
                "properties": {
                    "total_items": {
                        "description": "The total number of items available within the result set",
                        "type": "integer",
                        "format": "int64"
                    },
                    "items_per_page": {
                        "description": "How many items are provided per page",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
            "EntryGroup": {
                "title": "Entry Group",
                "description": "A named collection of entries sharing a status group, field layout, and category groups.",
                "properties": {
                    "id": {
                        "description": "Primary key",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Human-readable name",
                        "type": "string"
                    },
                    "handle": {
                        "description": "URL-safe identifier (unique)",
                        "type": "string"
                    },
                    "description": {
                        "description": "Optional description",
                        "type": "string",
                        "nullable": true
                    },
                    "sort_order": {
                        "description": "Display sort position",
                        "type": "integer",
                        "nullable": true
                    },
                    "status_group_id": {
                        "description": "ID of the associated status group",
                        "type": "integer",
                        "format": "int64"
                    },
                    "field_layout_id": {
                        "description": "ID of the associated field layout",
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    },
                    "entries_count": {
                        "description": "Total number of entries in this group (present when counted)",
                        "type": "integer",
                        "nullable": true
                    },
                    "entry_types_count": {
                        "description": "Total number of entry types in this group (present when counted)",
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}